refactor(catalog-popularity-freetext-resolve): extract enumerate query into @wxyc/database
Moves enumerateFreetextPairs (and its RawPair type) out of the
catalog-popularity-freetext-resolve job into shared/database/src/freetext-enumerate.ts,
mirroring the concerts-recompute.ts / concerts-artist-resolver recompute.ts shim pattern.
The job now re-exports the same symbols from @wxyc/database, and the integration spec
imports and executes the real function (via the built dist) instead of hand-maintaining
a SQL mirror, closing the drift risk the mirror existed to manage. The SQL-contract unit
tests moved from the job's test file to tests/unit/database/freetext-enumerate.test.ts,
which exercises the real module directly.

The package-level @wxyc/database jest mock gets a self-contained mock implementation of
enumerateFreetextPairs (not a re-export from source) since the real module's relative
./client.js import isn't caught by the existing client-mock moduleNameMapper entry and
would otherwise pull in the real DB client at mock-load time; the mock instead drives the
same db.transaction + two db.execute calls shape so the job's runResolve unit tests, which
sequence db.execute mock values, keep working unchanged.

No behavior change to the query itself.
