+++ jobs/catalog-popularity-freetext-resolve/job.ts
@@ -41,6 +41,8 @@ import {
   requireNonNegativeInt,
   requirePositiveInt,
   freetextPairKey,
+  enumerateFreetextPairs,
+  type RawPair,
 } from '@wxyc/database';
 import { bulkLookupMetadata, type BulkLookupItem, type BulkLookupResponse } from '@wxyc/lml-client';
 import * as Sentry from '@sentry/node';
@@ -133,16 +135,15 @@ export const STOP_BY_UTC_DEFAULT = '11:00';
 
 // -- Source query ------------------------------------------------------------
 
-/** A raw free-text pair as the DJ typed it, with a representative track title.
- * `song` is trimmed at the enumerate boundary (empty string when no usable
- * track exists for the pair) — it is the single place the "usable track?" rule
- * is applied, so every downstream consumer can treat a truthy `song` as ready
- * to send. */
-export interface RawPair {
-  artist: string;
-  album: string;
-  song: string;
-}
+/** `enumerateFreetextPairs` (and its `RawPair` shape) moved to
+ * `@wxyc/database` (BS#1799) so the `tests/integration` babel-jest harness —
+ * which can't import this job directly (no TS transform registered for
+ * `jest.config.json`) — can import the SAME statement the job runs instead of
+ * hand-duplicating a SQL mirror. Imported above (for use in `runResolve`
+ * below) and re-exported here so any existing import site of this job still
+ * resolves; see `shared/database/src/freetext-enumerate.ts` for the
+ * implementation and full rationale. */
+export { enumerateFreetextPairs, type RawPair };
 
 /** A normalized dedup key + the representative raw pair to send to LML. */
 export interface NormalizedPair {
@@ -152,70 +153,6 @@ export interface NormalizedPair {
   album: string;
   song: string;
 }
-
-/** SELECT one row per unlinked `(artist, album)` pair, carrying the pair's
- * MOST-PLAYED non-empty `track_title` as its representative track.
- *
- * The most-played track (not the alphabetically-first) is the album's
- * canonical track — an 'A…' bonus/intro title is an arbitrary, low-signal
- * representative that resolves to the wrong release (or no release) far more
- * often. Picking the modal track reduces both wrong-release matches and missed
- * matches; the A/B probe behind BS#1767 showed track-aware matching lifting the
- * match rate ~3.7x over album-only with zero regressions.
- *
- * Shape: an inner `GROUP BY (artist_name, album_title, track_title)` counts
- * plays per distinct track, then `DISTINCT ON (artist_name, album_title)` with
- * an `ORDER BY` that (1) prefers a non-empty track
- * (`btrim(coalesce(track_title, '')) = ''` sorts false-before-true, so
- * non-empty first), (2) then most-played (`play_count DESC`), (3) then a
- * deterministic `track_title ASC` tiebreak, keeps exactly the modal
- * representative per pair. Cardinality is UNCHANGED — still one row per
- * distinct `(artist, album)` pair; the GROUP BY only picks a better
- * representative track, it does not change which pairs are enumerated. There is
- * NO `track_title IS NOT NULL` filter — a pair whose plays are all track-less
- * still enumerates and resolves album-only, exactly as before this change.
- *
- * The `normalizePairs` determinism contract still holds: rows remain ordered by
- * `(artist_name, album_title)` first, so the first-encountered representative
- * per normalized key is stable across runs.
- *
- * The inner GROUP BY subquery measured ~17s on prod (within the raised
- * `statement_timeout`). Wrapped in `db.transaction` + `SET LOCAL
- * statement_timeout` because the `album_id IS NULL` partition isn't covered by
- * the metadata-drain partial indexes; the planner falls back to a scan that can
- * exceed the backend's default `statement_timeout`. `SET LOCAL` only scopes
- * inside an explicit transaction with the postgres-js driver. Mirrors
- * `album-level-backfill#enumeratePendingAlbumIds`. */
-export const enumerateFreetextPairs = async (timeoutMs: number = READ_TIMEOUT_DEFAULT): Promise<RawPair[]> => {
-  return await db.transaction(async (tx) => {
-    await tx.execute(sql.raw(`SET LOCAL statement_timeout = '${timeoutMs}ms'`));
-    const rows = (await tx.execute(sql`
-      SELECT DISTINCT ON ("artist_name", "album_title")
-             "artist_name", "album_title", "track_title"
-      FROM (
-        SELECT "artist_name", "album_title", "track_title", count(*) AS play_count
-        FROM "wxyc_schema"."flowsheet"
-        WHERE "entry_type" = 'track'
-          AND "album_id" IS NULL
-          AND "artist_name" IS NOT NULL
-          AND "album_title" IS NOT NULL
-        GROUP BY "artist_name", "album_title", "track_title"
-      ) g
-      ORDER BY "artist_name", "album_title",
-               (btrim(coalesce("track_title", '')) = '') ASC,
-               play_count DESC,
-               "track_title" ASC
-    `)) as unknown as Array<{ artist_name: string; album_title: string; track_title: string | null }>;
-    return rows.map((r) => ({
-      artist: String(r.artist_name),
-      album: String(r.album_title),
-      // Trim the representative track once, HERE, at the single "usable track?"
-      // boundary. Downstream (`buildBulkItems`) treats a truthy `song` as
-      // ready-to-send with no further trimming.
-      song: (r.track_title ?? '').trim(),
-    }));
-  });
-};
 
 /** Fold raw pairs into normalized dedup keys, keeping one representative raw
  * pair per key (the first encountered — `enumerateFreetextPairs` returns them
